INS-2216: future moved to hostnetwork
